JBDS-4074 Split requirements.json to windows ang mac version
Fix splits requirements.json to one per suppored platform:
1. win32 and darwin are real one used to produce installers;
2. linux one is just copy of original requirements.json
   to keep PR verification happy when running on linux hosts
