+++ requirements-darwin.json
@@ -1,0 +1,94 @@
+{
+  "cdk.zip": {
+    "name": "Red Hat Container Development Kit",
+    "description": "Developer Tools for Creating, Testing, and Distributing Red Hat Container-Based Applications",
+    "bundle": "yes",
+    "version": "2.3.GA",
+    "dmUrl": "http://cdk-builds.usersys.redhat.com/builds/nightly/21-Oct-2016/cdk.zip",
+    "url": "http://cdk-builds.usersys.redhat.com/builds/nightly/21-Oct-2016/cdk.zip",
+    "filename": "cdk_2.3.GA-20161021.zip",
+    "sha256sum": "a40d8d2c86bbac6a663cb8f22bca51dd9e552501ff4d469a5f68f5fce1590a43",
+    "virusTotalReport": "https://virustotal.com/en/file/a40d8d2c86bbac6a663cb8f22bca51dd9e552501ff4d469a5f68f5fce1590a43/analysis/",
+    "vendor": "Red Hat, Inc."
+  },
+  "rhel-vagrant-virtualbox.box": {
+    "name": "Red Hat Container Development Kit RHEL 7.2 Vagrant box for VirtualBox",
+    "description": "Developer Tools for Creating, Testing, and Distributing Red Hat Container-Based Applications",
+    "bundle": "yes",
+    "version": "7.2-30",
+    "dmUrl": "http://cdk-builds.usersys.redhat.com/builds/nightly/21-Oct-2016/rhel-7-cdk-vagrant-scratch-7.2.21102016-1.x86_64.vagrant-virtualbox.box",
+    "url": "http://cdk-builds.usersys.redhat.com/builds/nightly/21-Oct-2016/rhel-7-cdk-vagrant-scratch-7.2.21102016-1.x86_64.vagrant-virtualbox.box",
+    "filename": "rhel-7-cdk-vagrant-scratch-7.2.21102016-1.x86_64.vagrant-virtualbox.box",
+    "sha256sum": "b333bd1d3d2b7fbeadb8612a1c74137f1275fd8fca96fbf2a6580849d60f1962",
+    "vendor": "Red Hat, Inc."
+  },
+  "oc.zip": {
+    "name": "Openshift Origin",
+    "description": "The Open Source Container Application Platform",
+    "bundle": "yes",
+    "version": "1.3.1",
+    "url": "https://github.com/openshift/origin/releases/download/v1.3.1/openshift-origin-client-tools-v1.3.1-dad658de7465ba8a234a4fb40b5b446a45a4cee1-mac.zip",
+    "filename": "openshift-origin-client-tools-v1.3.1-dad658de7465ba8a234a4fb40b5b446a45a4cee1-mac.zip",
+    "sha256sum": "3c941f0634566142784503e0b06c9c733bc1fda75b7b8cffb71f688b4f400d03",
+    "virusTotalReport": "https://virustotal.com/en/file/3c941f0634566142784503e0b06c9c733bc1fda75b7b8cffb71f688b4f400d03/analysis/",
+    "vendor": "Red Hat, Inc."
+  },
+  "cygwin.exe": {
+    "name": "Cygwin",
+    "description": "A distribution of popular GNU and other Open Source tools running on Microsoft Windows",
+    "bundle": "always",
+    "version": "2.6.0",
+    "url": "https://cygwin.com/setup-x86_64.exe",
+    "filename": "cygwin_2.6.0.exe",
+    "sha256sum": "446b658bc1b8b6c7865474188cb4d7e9873003e2b6a9d74dcdfb7a3ff77e8634",
+    "virusTotalReport": "https://virustotal.com/en/file/446b658bc1b8b6c7865474188cb4d7e9873003e2b6a9d74dcdfb7a3ff77e8634/analysis/",
+    "vendor": "Community Project"
+  },
+  "jbds.jar": {
+    "name": "Red Hat JBoss Developer Studio",
+    "description": "A certified Eclipse-based integrated development environment (IDE)",
+    "bundle": "yes",
+    "version": "10.2.0.AM2",
+    "dmUrl": "http://www.qa.jboss.com/binaries/devstudio/10.0/staging/builds/devstudio-10.2.0.AM2-build-product/latest/all/devstudio-10.2.0.latest-installer-standalone.jar",
+    "url": "http://www.qa.jboss.com/binaries/devstudio/10.0/staging/builds/devstudio-10.2.0.AM2-build-product/latest/all/devstudio-10.2.0.latest-installer-standalone.jar",
+    "filename": "devstudio-10.2.0.latest-installer-standalone.jar",
+    "sha256sum": "http://www.qa.jboss.com/binaries/devstudio/10.0/staging/builds/devstudio-10.2.0.AM2-build-product/latest/all/devstudio-10.2.0.latest-installer-standalone.jar.sha256",
+    "vendor": "Red Hat, Inc."
+  },
+  "jdk.msi": {
+    "name": "Java Platform, Standard Edition",
+    "description": "An implementation of the Java Platform, Standard Edition (Java SE) by Oracle",
+    "bundle": "no",
+    "version": "1.8.0",
+    "prefix": "java-1.8.0-openjdk-1.8.0",
+    "dmUrl": "http://www.oracle.com/technetwork/java/javase/downloads/index.html",
+    "url": "http://www.oracle.com/technetwork/java/javase/downloads/index.html",
+    "filename": "jdk-8u111-macosx-x64.dmg",
+    "sha256sum": "",
+    "virusTotalReport": "",
+    "vendor": "Oracle, Inc."
+  },
+  "vagrant.msi": {
+    "name": "Vagrant",
+    "description": "A container provisioning tool",
+    "bundle": "no",
+    "version": "1.8.1",
+    "url": "https://releases.hashicorp.com/vagrant/1.8.1/vagrant_1.8.1.dmg",
+    "filename": "vagrant_1.8.1.dmg",
+    "sha256sum": "2cfdbeec9e40376e49dae9d9f27511896e3b296f0e24f8731339bb3d32c48c93",
+    "virusTotalReport": "https://virustotal.com/en/file/2cfdbeec9e40376e49dae9d9f27511896e3b296f0e24f8731339bb3d32c48c93/analysis/",
+    "vendor": "HashiCorp"
+  },
+  "virtualbox.exe": {
+    "name": "Oracle VirtualBox",
+    "description": "A virtualization software package developed by Oracle",
+    "bundle": "no",
+    "version": "5.0.26",
+    "revision": "108824",
+    "url": "http://download.virtualbox.org/virtualbox/5.0.26/VirtualBox-5.0.26-108824-OSX.dmg",
+    "filename": "virtualbox_5.0.26.dmg",
+    "sha256sum": "e8836a98adea9350917a41e754dfec4fe2df7c4a0224fd8beca72cbc5d778437",
+    "virusTotalReport": "https://virustotal.com/en/file/e8836a98adea9350917a41e754dfec4fe2df7c4a0224fd8beca72cbc5d778437/analysis/",
+    "vendor": "Oracle, Inc."
+  }
+}
