+++ requirements-win32.json
@@ -1,0 +1,113 @@
+  {
+  "cdk.zip": {
+    "name": "Red Hat Container Development Kit",
+    "description": "Developer Tools for Creating, Testing, and Distributing Red Hat Container-Based Applications",
+    "bundle": "yes",
+    "version": "2.3.GA",
+    "dmUrl": "http://cdk-builds.usersys.redhat.com/builds/nightly/21-Oct-2016/cdk.zip",
+    "url": "http://cdk-builds.usersys.redhat.com/builds/nightly/21-Oct-2016/cdk.zip",
+    "filename": "cdk_2.3.GA-20161021.zip",
+    "sha256sum": "a40d8d2c86bbac6a663cb8f22bca51dd9e552501ff4d469a5f68f5fce1590a43",
+    "virusTotalReport": "https://virustotal.com/en/file/a40d8d2c86bbac6a663cb8f22bca51dd9e552501ff4d469a5f68f5fce1590a43/analysis/",
+    "vendor": "Red Hat, Inc."
+  },
+  "rhel-vagrant-virtualbox.box": {
+    "name": "Red Hat Container Development Kit RHEL 7.2 Vagrant box for VirtualBox",
+    "description": "Developer Tools for Creating, Testing, and Distributing Red Hat Container-Based Applications",
+    "bundle": "yes",
+    "version": "7.2-30",
+    "dmUrl": "http://cdk-builds.usersys.redhat.com/builds/nightly/21-Oct-2016/rhel-7-cdk-vagrant-scratch-7.2.21102016-1.x86_64.vagrant-virtualbox.box",
+    "url": "http://cdk-builds.usersys.redhat.com/builds/nightly/21-Oct-2016/rhel-7-cdk-vagrant-scratch-7.2.21102016-1.x86_64.vagrant-virtualbox.box",
+    "filename": "rhel-7-cdk-vagrant-scratch-7.2.21102016-1.x86_64.vagrant-virtualbox.box",
+    "sha256sum": "b333bd1d3d2b7fbeadb8612a1c74137f1275fd8fca96fbf2a6580849d60f1962",
+    "vendor": "Red Hat, Inc."
+  },
+  "oc.zip": {
+    "name": "Openshift Origin",
+    "description": "The Open Source Container Application Platform",
+    "bundle": "yes",
+    "version": "1.3.1",
+    "url": "https://github.com/openshift/origin/releases/download/v1.3.1/openshift-origin-client-tools-v1.3.1-dad658de7465ba8a234a4fb40b5b446a45a4cee1-windows.zip",
+    "filename": "openshift-origin-client-tools-v1.3.1-dad658de7465ba8a234a4fb40b5b446a45a4cee1-windows.zip",
+    "sha256sum": "b90bc1249e7407717b0a0d7f92248ed6926ae0cd27d8fd038e054b866fa84baf",
+    "virusTotalReport": "https://virustotal.com/en/file/b90bc1249e7407717b0a0d7f92248ed6926ae0cd27d8fd038e054b866fa84baf/analysis/",
+    "vendor": "Red Hat, Inc."
+  },
+  "cygwin.exe": {
+    "name": "Cygwin",
+    "description": "A distribution of popular GNU and other Open Source tools running on Microsoft Windows",
+    "bundle": "always",
+    "version": "2.6.0",
+    "url": "https://cygwin.com/setup-x86_64.exe",
+    "filename": "cygwin_2.6.0.exe",
+    "sha256sum": "446b658bc1b8b6c7865474188cb4d7e9873003e2b6a9d74dcdfb7a3ff77e8634",
+    "virusTotalReport": "https://virustotal.com/en/file/446b658bc1b8b6c7865474188cb4d7e9873003e2b6a9d74dcdfb7a3ff77e8634/analysis/",
+    "vendor": "Community Project"
+  },
+  "jbds.jar": {
+    "name": "Red Hat JBoss Developer Studio",
+    "description": "A certified Eclipse-based integrated development environment (IDE)",
+    "bundle": "yes",
+    "version": "10.2.0.AM2",
+    "dmUrl": "http://www.qa.jboss.com/binaries/devstudio/10.0/staging/builds/devstudio-10.2.0.AM2-build-product/latest/all/devstudio-10.2.0.latest-installer-standalone.jar",
+    "url": "http://www.qa.jboss.com/binaries/devstudio/10.0/staging/builds/devstudio-10.2.0.AM2-build-product/latest/all/devstudio-10.2.0.latest-installer-standalone.jar",
+    "filename": "devstudio-10.2.0.latest-installer-standalone.jar",
+    "sha256sum": "http://www.qa.jboss.com/binaries/devstudio/10.0/staging/builds/devstudio-10.2.0.AM2-build-product/latest/all/devstudio-10.2.0.latest-installer-standalone.jar.sha256",
+    "vendor": "Red Hat, Inc."
+  },
+  "jdk.msi": {
+    "name": "OpenJDK",
+    "description": "A free and open source implementation of the Java Platform, Standard Edition (Java SE)",
+    "bundle": "yes",
+    "version": "1.8.0.111",
+    "dmUrl": "https://developers.redhat.com/download-manager/jdf/file/java-1.8.0-openjdk-1.8.0.111-1.b15.redhat.windows.x86_64.msi?workflow=direct",
+    "url": "http://download-node-02.eng.bos.redhat.com/brewroot/packages/openjdk8-win/1.8.0.111/1.b15/win/java-1.8.0-openjdk-1.8.0.111-1.b15.redhat.windows.x86_64.msi",
+    "filename": "java-1.8.0-openjdk-1.8.0.111-1.b15.redhat.windows.x86_64.msi",
+    "sha256sum": "fc4c77412446d6bbf6bf91d76d4bb5cf5d1e94dc1e932f524796e3f8a9789ccc",
+    "virusTotalReport": "https://virustotal.com/en/file/fc4c77412446d6bbf6bf91d76d4bb5cf5d1e94dc1e932f524796e3f8a9789ccc/analysis/",
+    "vendor": "Red Hat, Inc."
+  },
+  "vagrant.msi": {
+    "name": "Vagrant",
+    "description": "A container provisioning tool",
+    "bundle": "no",
+    "version": "1.8.1",
+    "url": "https://developers.redhat.com/redirect/to/vagrant-1.8.1.download",
+    "filename": "vagrant_1.8.1.msi",
+    "sha256sum": "6468cc140f6f236c3ac4bc0bbb1cc2177114249bc344e0aba8c826eae9cc6eb2",
+    "virusTotalReport": "https://virustotal.com/en/file/6468cc140f6f236c3ac4bc0bbb1cc2177114249bc344e0aba8c826eae9cc6eb2/analysis/",
+    "vendor": "HashiCorp"
+  },
+  "virtualbox.exe": {
+    "name": "Oracle VirtualBox",
+    "description": "A virtualization software package developed by Oracle",
+    "bundle": "no",
+    "version": "5.0.26",
+    "revision": "108824",
+    "url": "http://developers.redhat.com/redirect/to/virtualbox-5.0.26.download",
+    "filename": "virtualbox_5.0.26.exe",
+    "sha256sum": "7195c5c71add512e8ac97ffc995f5a809d239ac550e20017c1d6d4766a0ced1c",
+    "virusTotalReport": "https://virustotal.com/en/file/7195c5c71add512e8ac97ffc995f5a809d239ac550e20017c1d6d4766a0ced1c/analysis/",
+    "vendor": "Oracle, Inc."
+  },
+  "7zip.zip": {
+    "name": "7zip",
+    "description": "7Zip Command Line Version",
+    "bundle": "tools",
+    "version": "9.20",
+    "url": "https://downloads.sourceforge.net/project/sevenzip/7-Zip/9.20/7za920.zip?r=https%3A%2F%2Fsourceforge.net%2Fprojects%2Fsevenzip%2Ffiles%2F7-Zip%2F9.20%2F",
+    "filename": "7za920.zip",
+    "sha256sum": "2a3afe19c180f8373fa02ff00254d5394fec0349f5804e0ad2f6067854ff28ac",
+    "vendor": "Igor Pavlov"
+  },
+  "7zip-extra.zip": {
+    "name": "7zipextra",
+    "description": "7Zip Extra: SFXs for installers",
+    "bundle": "tools",
+    "version": "9.20",
+    "url": "https://downloads.sourceforge.net/project/sevenzip/7-Zip/9.20/7z920_extra.7z?r=https%3A%2F%2Fsourceforge.net%2Fprojects%2Fsevenzip%2Ffiles%2F7-Zip%2F9.20%2F",
+    "filename": "7z920_extra.7z",
+    "sha256sum": "d65a1d1a050e4fd7912c18468954f64db824a1e1b621235a153d010beae14757",
+    "vendor": "Igor Pavlov"
+  }
+}
